Merging

+++ Interfaces/WebStoreServices/IItemService.cs
@@ -130,7 +130,7 @@ namespace MPC.Interfaces.WebStoreServices
 
         Item GetItemByOrderItemID(long ItemID, long OrderID);
 
-        long AddInquiryAndItems(Inquiry Inquiry, List<InquiryItem> InquiryItems);
+        int AddInquiryAndItems(Inquiry Inquiry, List<InquiryItem> InquiryItems);
         void AddInquiryAttachments(List<InquiryAttachment> InquiryAttachments);
         /// <summary>
         /// get item section
